add logging for dnsmasq messages
By default dnsmasq logs to DAMEON which normally hits /var/log/syslog or /var/log/messages
This isn't the case with the containers.

I added an override to log to /var/log/neutron/neutron-dnsmasq.log \
			aka /openstack/log/*neutron_agents*/neutron-dnsmasq.log

This should automatically get pulled in and logstash'd when the play is rerun based on the code below.
Someone please confirm that no additional steps are needed to pull in this extra log.

https://github.com/rcbops/ansible-lxc-rpc/blob/master/rpc_deployment/roles/rsyslog_config/templates/genlog.py#L26-28

"""
for root, dirnames, filenames in os.walk('/openstack/log'):
        for filename in fnmatch.filter(filenames, '*.log'):
                matches.append(os.path.join(root, filename))
"""